Add block scale dequantize CPU reference implementation (#6582)
## Summary

- Adds `CpuFpReferenceBlockScaleDequantize` utility for computing the
CPU reference output of block scale dequantize operations (MX types:
fp4, fp6, fp8 with fp8_e8m0 scale)
- Adds `BlockScaleDequantizeSignatureKey` and `BlockScaleDequantizePlan`
to the `CpuReferenceGraphExecutor` plan registry
- Extends `CpuReferenceGraphExecutor` to dispatch block scale dequantize
nodes
- Adds integration test (`TestCpuReferenceGraphExecutor`) covering the
full execute path, plus unit tests for the plan and signature key
- Extends `FlatbufferDatatypeMapping` with MX data types (fp8_e4m3,
fp8_e5m2, fp8_e8m0, fp4_e2m1, fp6_e2m3, fp6_e3m2)
- Updates `FlatbufferGraphTestUtils` with separate `scaleDataType` and
`outputDataType` parameters in `createValidBlockScaleDequantizeGraph`
(addresses reviewer feedback)

## Test plan

- [x] `ninja check` in `projects/hipdnn/build` — all 9 test suites pass,
0 failures

+++ projects/hipdnn/test_sdk/include/hipdnn_test_sdk/utilities/CpuFpReferenceValidation.hpp
@@ -67,7 +67,7 @@ public:
                 return result.load(std::memory_order_relaxed);
             }
 
-            auto absDiff = static_cast<float>(fabs(implValue - refValue));
+            auto absDiff = fabs(static_cast<float>(implValue) - static_cast<float>(refValue));
             auto threshold
                 = _absoluteTolerance + _relativeTolerance * fabs(static_cast<float>(refValue));
